+++ lexiang_notice/src/main/java/top/lexiang/notice/controller/NoticeController.java
@@ -2,18 +2,20 @@ package top.lexiang.notice.controller;
 
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
+import top.lexiang.common.controller.BaseController;
 import top.lexiang.common.entity.PageResult;
 import top.lexiang.common.entity.Result;
 import top.lexiang.common.entity.ResultCode;
-import top.lexiang.notice.entity.Notice;
-import top.lexiang.notice.entity.NoticeFresh;
+import top.lexiang.entity.notice.Notice;
+import top.lexiang.entity.notice.NoticeFresh;
 import top.lexiang.notice.service.NoticeService;
 
 @CrossOrigin
 @RestController
 @RequestMapping("/notice")
-public class NoticeController {
+public class NoticeController extends BaseController {
 
     @Autowired
     private NoticeService noticeService;
@@ -45,6 +47,8 @@ public class NoticeController {
     //3.新增通知
     @PostMapping()
     public Result save(@RequestBody Notice notice) {
+        //设置进行操作用户的id  获取用户的角色 设置消息是系统消息（system）还是用户消息（uesr)
+        notice.setOperatorId(this.userId); //操作者 A对B进行点赞 该参数是设置A的值
         noticeService.save(notice);
         return new Result(ResultCode.SUCCESS);
     }
@@ -52,6 +56,12 @@ public class NoticeController {
     //4.修改通知
     @PutMapping()
     public Result updateById(@RequestBody Notice notice) {
+        //判断该通知是否为本人所发送
+        Notice notice1 = noticeService.selectById(notice.getId());
+        if (!this.userId.equals(notice1.getOperatorId()))
+        {
+            return new Result(ResultCode.UNAUTHORISE);
+        }
         int count = noticeService.updateById(notice);
         if (count < 1) {
             return new Result(ResultCode.TARGETISEMPTY);
@@ -60,11 +70,11 @@ public class NoticeController {
     }
 
     //5.根据用户id查询该用户的待推送消息
-    @GetMapping("/fresh/{userId}/{page}/{size}")
-    public Result freshPage(@PathVariable("userId") String userId,
-                            @PathVariable("page") Integer page,
+    @GetMapping("/fresh/{page}/{size}")
+    public Result freshPage(@PathVariable("page") Integer page,
                             @PathVariable("size") Integer size)  {
-        Page<NoticeFresh> pageData = noticeService.freshPage(userId, page, size);
+
+        Page<NoticeFresh> pageData = noticeService.freshPage(this.userId, page, size);
 
         if (pageData == null) {
             return new Result(ResultCode.VALUEISEMPTY);
@@ -86,6 +96,11 @@ public class NoticeController {
     //7.删除消息
     @DeleteMapping("/{id}")
     public Result delete(@PathVariable("id") String id) {
+        Notice notice = noticeService.selectById(id);
+        //删除的消息是否属于该用户 且该用户不为管理员
+        if (!this.userId.equals(notice.getOperatorId()) && StringUtils.isEmpty(this.roleId)) {
+            return new Result(ResultCode.UNAUTHORISE);
+        }
         int count = noticeService.delete(id);
         if (count < 1) {
             return new Result(ResultCode.TARGETISEMPTY);
